Refactor GlueShader.cpp (NFC)
GlueShader.cpp contains the implementation of 3 classes, and I want to
add another type of glue shader.  This is too much for a single file, so
I want to split this file up and move each cleass into is own file.
